Bind Deform360 audit to physical source directory
